+++ packages/astro/src/assets/fonts/types.ts
@@ -5,7 +5,7 @@ import type {
 	localFontFamilySchema,
 	remoteFontFamilySchema,
 } from './config.js';
-import type { FONT_TYPES } from './constants.js';
+import type { FONT_TYPES, GENERIC_FALLBACK_NAMES } from './constants.js';
 import type { Font } from '@capsizecss/unpack';
 
 export type AstroFontProvider = z.infer<typeof fontProviderSchema>;
@@ -65,3 +65,5 @@ export type FontFaceMetrics = Pick<
 	Font,
 	'ascent' | 'descent' | 'lineGap' | 'unitsPerEm' | 'xWidthAvg'
 >;
+
+export type GenericFallbackName = (typeof GENERIC_FALLBACK_NAMES)[number];
